pyproject.toml
[tool.poetry]
name = "auto-grade"
version = "0.1.0"
description = "Automatically give feedback in Person ActiveLearn"
authors = ["Samuel Searles-Bryant <devel@samueljsb.co.uk>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^3.141.0"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"
flake8 = "^3.7.9"
flake8-docstrings = "^1.5.0"
mypy = "^0.770"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
